refac: move the references to separate component
